requirements.txt
dj-database-url==0.5.0
Django==2.2.4
django-betterforms==1.2
django-heroku==0.3.1
djangorestframework==3.10.3
gunicorn==19.9.0
Pillow==6.2.0
psycopg2==2.8.3
python-decouple==3.1
pytz==2019.3
sqlparse==0.3.0
whitenoise==4.1.4
